Add source to Xcode project
