+++ Brain/Brain.h
@@ -1,0 +1,19 @@
+//
+//  Brain.h
+//  Brain
+//
+//  Created by Douglas Bumby on 2015-05-31.
+//  Copyright (c) 2015 Cosmic Labs. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+
+//! Project version number for Brain.
+FOUNDATION_EXPORT double BrainVersionNumber;
+
+//! Project version string for Brain.
+FOUNDATION_EXPORT const unsigned char BrainVersionString[];
+
+// In this header, you should import all the public headers of your framework using statements like #import <Brain/PublicHeader.h>
+
+
