+++ Brain/NeuralNetwork.swift
@@ -1,8 +1,9 @@
 //
-//  main.swift
-//  swiftex1
+//  NeuralNetwork.swift
+//  Brain
 //
-//  Created by V Lall on 6/8/14
+//  Created by Vishal on 2014-06-08.
+//  Copyright (c) 2015 Vishal. All rights reserved.
 //
 
 import Foundation
@@ -22,12 +23,14 @@ func *& (fill: Array<Double>, I: NSInteger) -> Array<Double>{
             m.append(fill[index])
         }
     }
+    
     return m
 }
 
 func randomFunc(a: Double, b:Double) -> (Double) {
     var randNum = arc4random_uniform(100)/100
     var output = (b-a)*Double(randNum) + (a)
+    
     return output
 }
 
@@ -88,29 +91,31 @@ class NN {
                 self.wi[i][j]=randomFunc(-0.2, 0.2)
             }
         }
+        
         for j in 0...(self.nh-1){
             for k in 0...(self.no-1){
                 self.wo[j][k] = randomFunc(-2.0, 2.0)
             }
         }
-
         
         // last change in weights for momentum
         self.ci = makeMatrix(self.ni, self.nh)
         self.co = makeMatrix(self.nh, self.no)
 
     }
+    
     func update(inputs:Array<Double>) -> (Array<Double>) {
         if (inputs.count != self.ni-1){
             println("wrong number of inputs")
         }
-       // input activations
-/*
-        println(inputs)
-        println(self.ai)
-        println(self.ni)
-        println(inputs.count)
-        println(self.ni-1)*/
+        
+        // input activations
+        // println(inputs)
+        // println(self.ai)
+        // println(self.ni)
+        // println(inputs.count)
+        // println(self.ni-1)
+        
         for i in 0..(self.ni-1){
         //self.ai[i] = sigmoid(inputs[i])
             self.ai[i] = inputs[i]
@@ -121,17 +126,21 @@ class NN {
             for i in 1...(self.ni-1){
                 sum = sum + self.ai[i] * self.wi[i][j]
             }
+            
             self.ah[j] = sigmoid(sum)
             
         }
+        
         // output activations
         for k in 1...(self.no-1){
             var sum = 0.0
             for j in 1...(self.nh-1){
                 sum = sum + self.ah[j] * self.wo[j][k]
             }
+            
             self.ao[k] = sigmoid(sum)
         }
+        
         return self.ao
     }
     
@@ -139,21 +148,25 @@ class NN {
         if targets.count != self.no{
             println("wrong number of target values")
         }
+        
         // calculate error terms for output
         var output_deltas = [0.0] *& self.no
         for k in 0..(self.no){
             var error = targets[k]-self.ao[k]
             output_deltas[k] = dsigmoid(self.ao[k]) * error
         }
+        
         // calculate error terms for hidden
-       var hidden_deltas = [0.0] *& self.nh
+        var hidden_deltas = [0.0] *& self.nh
         for j in 0..(self.nh){
             var error = 0.0
             for k in 0..(self.no){
                 error = error + output_deltas[k]*self.wo[j][k]
             }
+            
             hidden_deltas[j] = dsigmoid(self.ah[j]) * error
         }
+        
         // update output weights
         for j in 0..(self.nh){
             for k in 0..(self.no){
@@ -163,6 +176,7 @@ class NN {
                 //print N*change, M*self.co[j][k]
             }
         }
+        
         // update input weights
         for i in 0..(self.ni){
             for j in 0..(self.nh){
@@ -171,11 +185,13 @@ class NN {
                 self.ci[i][j] = change
             }
         }
+        
         // calculate error
         var error = 0.0
         for k in 0...(targets.count){
             error = error + 0.5*(targets[k]-self.ao[k])**2
         }
+        
         return error
     }
     
@@ -184,17 +200,20 @@ class NN {
          //   println("\(patterns[p][0]) ->  \(self.update(patterns[p][0]))")
         }
     }
+    
     func weights()->(){
         print("Input weights:")
         for i in 0..(self.ni){
             println(self.wi[i])
             println("Output weights:")
         }
+        
         for j in 0..(self.nh){
             print(self.wo[j])
         }
     
     }
+    
     func train(patterns:Array<Array<Array<Double>>>, iterations:NSInteger=1000, N:Double=0.5, M:Double=0.1){
     // N: learning rate
     // M: momentum factor
@@ -206,14 +225,16 @@ class NN {
                 self.update(inputs)
                // error = error + self.backPropagate(targets, N: N, M: M)
             }
+            
             if i % 100 == 0{
                 println("error \(error)")
             }
         }
     }
 }
+
 func demo()->(){
-//Teach network XOR function
+    //Teach network XOR function
     var pat = Array<Array<Array<Double>>>()
 
     pat = [
@@ -231,9 +252,8 @@ func demo()->(){
     n.test(pat)
     }
 
-//let myFirstNN = NN(ni: 10,nh: 10,no: 10)
-//var x = [2.0]*&4
-println(demo())
-
+    // let myFirstNN = NN(ni: 10,nh: 10,no: 10)
+    // var x = [2.0]*&4
+    println(demo())
