+++ Brain/statistics/kalman.swift
@@ -1,19 +1,21 @@
-// Kalman Filter development underway
 //
-//  kalman.swift
-//  swiftex1
+//  Kalman.swift
+//  Brain
 //
-//  Created by V Lall on 6/27/14.
-//  Copyright (c) 2014 swiftintel. All rights reserved.
+//  Created by Vishal on 2014-06-27.
+//  Copyright (c) 2015 Vishal. All rights reserved.
+//
+//  Kalman Filter development underway
+//  Converted form python kalmanfilter, needs to be fixed.
+//  Make a dot product function
 //
 
-//Converted form python kalmanfilter, needs to be fixed.
-//Make a dot product function
 import Foundation
 
 func kf_predict(X, P, A, Q, B, U)->(){
     X = dot(A, X) + dot(B, U)
     P = dot(A, dot(P, A.T)) + Q
+    
     return(X,P)
 }
 
@@ -24,6 +26,7 @@ func kf_update(X, P, Y, H, R)->(){
     X = X + dot(K, (Y-IM))
     P = P - dot(K, dot(IS, K.T))
     LH = gauss_pdf(Y, IM, IS)
+    
     return (X,P,K,IM,IS,LH)
 }
 
@@ -42,5 +45,6 @@ func gauss_pdf(X, M, S)->(){
     E = 0.5 * dot(DX.T, dot(inv(S), DX))
     E = E + 0.5 * M.shape()[0] * log(2 * pi) + 0.5 * log(det(S))
     P = exp(-E)
+    
     return (P[0],E[0])
 }
